make tag company id configurable
Main reason for this was that previously the company id used for
filtering scan results was set only in one place for empty configuration
struct. That struct could corrupt in NVS flash (mostly after flashing a
new firmware) and there was no way to set it back to real value. Wifi
hotspot page has option for this id now.